Fixed a bug where files weren't detected as deleted when a file with the same properties was deleted before.
Former-commit-id: fa9b618f898fb1b60b4e12afc32b49ce6733d25c
Former-commit-id: 89c6c9f4d5bd7ffb3091988da0db6d57518e0903

+++ nodejs/Upload.js
@@ -62,6 +62,14 @@ Upload.prototype.deleteUpload = function() {
 };
 
 /**
+ * Returns if the Upload is deleted.
+ * @returns {boolean}
+ */
+Upload.prototype.isDeleted = function() {
+    return this.deleted;
+};
+
+/**
  * Adds the deletedByNickname to the Upload.
  * @param {string} deletedByNickname
  */
@@ -93,7 +101,7 @@ function addDeletedBy(channelID, filename, deletedByNickname, deletedByID) {
         shortFilename = filename.substr(filename.indexOf("\\/") + 1);
     }
     for (var i = 0; i < UploadList.length; i++) {
-        if (UploadList[i].getChannelID() == channelID && UploadList[i].getFilename() == shortFilename) {
+        if (UploadList[i].getChannelID() == channelID && UploadList[i].getFilename() == shortFilename && !UploadList[i].isDeleted()) {
             UploadList[i].addDeletedByNickname(deletedByNickname);
             UploadList[i].addDeletedByID(deletedByID);
             UploadList[i].deleteUpload();
